[Feature/cb-10] Allgemeine Commands (#3)
This PR adds common commands to the citybuild plugin according to this
[ticket](https://quantraxnet.youtrack.cloud/issue/CB-10/Allgemeine-Commands)
by @ByTRYO on YouTrack and some adjustments for developers.
Since I don't have a 2nd account for minecraft, all commands relating to
teleportation have been outsourced to another ticket. It can be found
[here](https://quantraxnet.youtrack.cloud/issue/CB-23/Teleport-Commands).

Also, messages have been migrated to a database but are still have to
created by hand.
___

### Minecraft
It contains the following citybuild features:
- Farmwelt Command <sup>1
- Nether Command <sup>1
- Heal Command ²
- Feed Command ²
- Day Command ²
- Invsee Command ²
- Sign Command ²

1 = without permissions
2 = with permission


### Developer
It contains the following features for developers:
- ItemBuilder (likely to be removed when the CoreAPI has a stable v2
release)
- Simple custom gui framework
- Auto-CleanUp and auto-updater for messages from `messages.toml` to
database


For future implemententions of clickable guis or text inputs, the
libraries
[InventoryGUI](https://wiki.phoenix616.dev/library/inventorygui/start)
and [AnvilGUI ](https://github.com/WesJD/AnvilGUI)have been added as
dependencies.
___
### Issue-Tracking
Please report any issues or other requests
[here](https://quantraxnet.youtrack.cloud/issues/CB).
___

### Notes
Please note that messages still have to be created by hand in a file
called `messages.toml`. The _auto-updater_ keeps track of those messages
and automatically inserts new entries into the database. The
_auto-cleanUp_ validates the serverside registered messages and removes
them if they don't exists in the messages file.

Please note that the underlaying database structure is NOT the current
structure that is applied to the production database. The testing
structure is divided into more context specific schemas and thus these
schema names are prefixed to the table names.
The simplefied version can be found below. 
To connect to your local database, create a credentials.env file in your
resources directory and paste
[this](https://discord.com/channels/491680017072390145/1070403827460739214/1208203084120199218)
into the file (must be ignored by .gitignore)

```mermaid
classDiagram
direction BT
class message {
   varchar(20) plugin
   text de
   text en
   varchar(60) i18n
}
```